move peer manager back into peer.go
